+++ src/main/java/crazypants/enderio/ModObject.java
@@ -122,7 +122,6 @@ public enum ModObject {
 
     itemEnderFood,
 
-    itemInfiniteTravelStaff,
     itemTeleportStaff;
 
     public final String unlocalisedName;
